refactor: split test class, add XML docs, clean up legacy naming
- Split Given_McpProxy into 5 test classes matching production classes
  (Given_HealthService, Given_ToolCacheFile, Given_McpUpstreamClient,
  Given_DevServerMonitor, Given_McpStdioServer)
- Add XML doc comments on test classes with <see cref> to tested class
- Add MSTest [Description] on non-trivial test methods
- Rename mcpClientProxy parameter to mcpUpstreamClient in all classes
- Replace "MCP proxy" with "MCP bridge" / "stdio server" in docs and logs
- Remove Phase 1b/1c-1 references from code comments
- Add [Retry(2)] on intermittent DevServer process tests
